add auth

+++ graphql/resolvers/booking.js
@@ -3,20 +3,29 @@ const Event = require('../../models/event');
 const { transformEvent, transformBooking } = require('./merge');
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (a, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
     const bookings = await Booking.find();
     return bookings.map((booking) => transformBooking(booking.toObject()));
   },
-  bookEvent: async (a) => {
+  bookEvent: async (a, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
     const fetchedEvent = await Event.findOne({ _id: a.eventId });
     const booking = new Booking({
-      user: '603f19e6924be91c85348d12',
+      user: req.userId,
       event: fetchedEvent
     });
     const result = await booking.save();
     return transformBooking(result.toObject());
   },
-  cancelBooking: async (a) => {
+  cancelBooking: async (a, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
     const fetchedBooking = await Booking.findById(a.bookingId);
     await Booking.deleteOne({ _id: a.bookingId });
     return transformEvent(fetchedBooking.toObject().event);
